refactor(desktop): extract move pane logic to separate actions file
Move `movePaneToTab` and `movePaneToNewTab` logic to
`stores/tabs/actions/move-pane.ts` to keep the main store file focused.

The store now delegates to pure functions that compute the new state,
making the logic easier to test and maintain.
